LayoutTests/editing/input/ios/invalidate-text-entry-context-when-clearing-text-area.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true ] -->
<html>
<meta name="viewport" content="width=device-width, initial-scale=1">
<meta charset="utf-8">
<head>
<script src="../../../resources/js-test.js"></script>
<script src="../../../resources/ui-helper.js"></script>
<style>
body {
    margin: 0;
    font-family: system-ui;
    line-height: 150%;
}

textarea {
    border: 1px solid tomato;
    box-sizing: border-box;
    outline: none;
    font-size: 18px;
    width: 300px;
}
</style>
<script>
jsTestIsAsync = true;

addEventListener("load", async () => {
    description("This test requires WebKitTestRunner.");

    const textarea = document.querySelector("textarea");
    await UIHelper.activateElementAndWaitForInputSession(textarea);

    for (let character of [..."hello"]) {
        await UIHelper.callFunctionAndWaitForEvent(async () => {
            await UIHelper.typeCharacter(character);
            await UIHelper.ensurePresentationUpdate();
        }, textarea, "keyup");
    }

    updateCountBefore = await UIHelper.keyboardUpdateForChangedSelectionCount();
    textarea.value = "";
    await UIHelper.ensurePresentationUpdate();
    updateCountAfter = await UIHelper.keyboardUpdateForChangedSelectionCount();

    shouldNotBe("updateCountBefore", "updateCountAfter");

    textarea.blur();
    await UIHelper.waitForKeyboardToHide();
    finishJSTest();
});
</script>
</head>
<body>
    <textarea></textarea>
</body>
</html>
