LayoutTests/editing/input/ios/invalidate-text-entry-context-when-clearing-text-area-expected.txt

This test requires WebKitTestRunner.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS updateCountBefore is not updateCountAfter
PASS successfullyParsed is true

TEST COMPLETE
